Convert openssl_privatekey_pipe to action plugin.
