+++ plugins/modules/openssl_privatekey_pipe.py
@@ -16,7 +16,8 @@ version_added: 1.2.0
 description:
     - This module allows one to (re)generate OpenSSL private keys without disk access.
     - This allows to read and write keys to vaults without having to write intermediate versions to disk.
-    - Make sure to not write the result of this module into logs or to the console, as it contains private key data!
+    - Make sure to not write the result of this module into logs or to the console, as it contains private key data! Use the I(no_log) task option to be sure.
+    - Note that this module is implemented as an L(action plugin,https://docs.ansible.com/ansible/latest/plugins/action.html) and will always be executed on the controller.
 author:
     - Yanis Guenane (@Spredzy)
     - Felix Fontein (@felixfontein)
@@ -97,106 +98,3 @@ privatekey:
     returned: always
     type: str
 '''
-
-import base64
-
-from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils._text import to_native, to_bytes
-
-from ansible_collections.community.crypto.plugins.module_utils.crypto.basic import (
-    OpenSSLObjectError,
-)
-
-from ansible_collections.community.crypto.plugins.module_utils.crypto.module_backends.privatekey import (
-    select_backend,
-    get_privatekey_argument_spec,
-)
-
-
-class PrivateKeyModule(object):
-    def __init__(self, module, module_backend):
-        self.module = module
-        self.module_backend = module_backend
-        self.check_mode = module.check_mode
-        self.changed = False
-
-        if module.params['content'] is not None:
-            if module.params['content_base64']:
-                try:
-                    data = base64.b64decode(module.params['content'])
-                except Exception as e:
-                    module.fail_json(msg='Cannot decode Base64 encoded data: {0}'.format(e))
-            else:
-                data = to_bytes(module.params['content'])
-            module_backend.set_existing(data)
-
-    def generate(self, module):
-        """Generate a keypair."""
-
-        if self.module_backend.needs_regeneration():
-            # Regenerate
-            if not self.check_mode:
-                self.module_backend.generate_private_key()
-                privatekey_data = self.module_backend.get_private_key_data()
-                self.privatekey_bytes = privatekey_data
-            self.changed = True
-        elif self.module_backend.needs_conversion():
-            # Convert
-            if not self.check_mode:
-                self.module_backend.convert_private_key()
-                privatekey_data = self.module_backend.get_private_key_data()
-                self.privatekey_bytes = privatekey_data
-            self.changed = True
-
-    def dump(self):
-        """Serialize the object into a dictionary."""
-        result = self.module_backend.dump(include_key=True)
-        result['changed'] = self.changed
-        return result
-
-
-def main():
-
-    argument_spec, required_together, required_if = get_privatekey_argument_spec()
-    argument_spec.update(dict(
-        content=dict(type='str', no_log=True),
-        content_base64=dict(type='bool', default=False),
-    ))
-    module = AnsibleModule(
-        argument_spec=argument_spec,
-        supports_check_mode=True,
-        required_together=required_together + [],
-        required_if=required_if + [],
-    )
-    # This could also be passed to the AnsibleModule() construction, but will be
-    # overwritten when the module's options are processed (Ansible always passes
-    # _ansible_no_log which overwrites it). We later have to remove the value of
-    # content from the `module.no_log_values` set, and we want to avoid accidental
-    # logging later on.
-    module.no_log = True
-
-    backend, module_backend = select_backend(
-        module=module,
-        backend=module.params['select_crypto_backend'],
-    )
-
-    try:
-        private_key = PrivateKeyModule(module, module_backend)
-        private_key.generate(module)
-        result = private_key.dump()
-        # In case changed=False, the module's input (`content`) is returned as `privatekey`.
-        # Since `content` is no_log=True, `privatekey`'s value will get replaced by
-        # ANSIBLE_NO_LOG_VALUE. To avoid this, we remove the value of `content` from
-        # module.no_log_values. Since we explicitly set `module.no_log = True` above, this
-        # should be safe.
-        try:
-            module.no_log_values.remove(module.params['content'])
-        except KeyError:
-            pass
-        module.exit_json(**result)
-    except OpenSSLObjectError as exc:
-        module.fail_json(msg=to_native(exc))
-
-
-if __name__ == '__main__':
-    main()
